kubecon-2022/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.8.3
    # via
    #   aiohttp-cors
    #   ray
aiohttp-cors==0.7.0
    # via ray
aiosignal==1.2.0
    # via
    #   aiohttp
    #   ray
arrow==1.2.3
    # via jinja2-time
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   ray
    #   visions
binaryornot==0.4.4
    # via cookiecutter
blessed==1.19.1
    # via gpustat
cachetools==5.2.0
    # via google-auth
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
chardet==5.0.0
    # via binaryornot
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
click==8.0.4
    # via
    #   cookiecutter
    #   flytekit
    #   ray
cloudpickle==2.2.0
    # via flytekit
colorful==0.5.4
    # via ray
cookiecutter==2.1.1
    # via flytekit
croniter==1.3.7
    # via flytekit
cryptography==38.0.1
    # via pyopenssl
cycler==0.11.0
    # via matplotlib
dataclasses-json==0.5.7
    # via
    #   -r requirements.in
    #   flytekit
decorator==5.1.1
    # via retry
deprecated==1.2.13
    # via flytekit
diskcache==5.4.0
    # via flytekit
distlib==0.3.6
    # via virtualenv
docker==6.0.0
    # via flytekit
docker-image-py==0.1.12
    # via flytekit
docstring-parser==0.15
    # via flytekit
filelock==3.8.0
    # via
    #   ray
    #   virtualenv
flyteidl==1.1.22
    # via
    #   flytekit
    #   flytekitplugins-ray
flytekit==1.2.3
    # via
    #   -r requirements.in
    #   flytekitplugins-deck-standard
    #   flytekitplugins-ray
    #   flytekitplugins-spark
flytekitplugins-deck-standard==1.2.3
    # via -r requirements.in
flytekitplugins-ray==1.2.3
    # via -r requirements.in
flytekitplugins-spark==1.2.3
    # via -r requirements.in
fonttools==4.38.0
    # via matplotlib
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
    #   ray
google-api-core==2.10.2
    # via opencensus
google-auth==2.13.0
    # via google-api-core
googleapis-common-protos==1.56.4
    # via
    #   flyteidl
    #   google-api-core
    #   grpcio-status
gpustat==1.0.0
    # via ray
grpcio==1.43.0
    # via
    #   flytekit
    #   grpcio-status
    #   ray
grpcio-status==1.43.0
    # via
    #   -r requirements.in
    #   flytekit
htmlmin==0.1.12
    # via pandas-profiling
idna==3.4
    # via
    #   requests
    #   yarl
imagehash==4.3.1
    # via visions
importlib-metadata==5.0.0
    # via
    #   flytekit
    #   keyring
    #   markdown
jaraco-classes==3.2.3
    # via keyring
jinja2==3.1.2
    # via
    #   cookiecutter
    #   jinja2-time
    #   pandas-profiling
jinja2-time==0.2.0
    # via cookiecutter
joblib==1.2.0
    # via
    #   flytekit
    #   phik
    #   scikit-learn
jsonschema==4.16.0
    # via ray
keyring==23.9.3
    # via flytekit
kiwisolver==1.4.4
    # via matplotlib
markdown==3.4.1
    # via flytekitplugins-deck-standard
markupsafe==2.1.1
    # via jinja2
marshmallow==3.18.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
    #   marshmallow-jsonschema
marshmallow-enum==1.5.1
    # via dataclasses-json
marshmallow-jsonschema==0.13.0
    # via flytekit
matplotlib==3.5.3
    # via
    #   missingno
    #   pandas-profiling
    #   phik
    #   seaborn
missingno==0.5.1
    # via pandas-profiling
more-itertools==9.0.0
    # via jaraco-classes
msgpack==1.0.4
    # via ray
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
multimethod==1.9
    # via
    #   pandas-profiling
    #   visions
mypy-extensions==0.4.3
    # via typing-inspect
natsort==8.2.0
    # via flytekit
networkx==2.8.7
    # via visions
numpy==1.23.4
    # via
    #   imagehash
    #   matplotlib
    #   missingno
    #   palmerpenguins
    #   pandas
    #   pandas-profiling
    #   patsy
    #   phik
    #   pyarrow
    #   pywavelets
    #   ray
    #   scikit-learn
    #   scipy
    #   seaborn
    #   statsmodels
    #   visions
nvidia-ml-py==11.495.46
    # via gpustat
opencensus==0.11.0
    # via ray
opencensus-context==0.1.3
    # via opencensus
packaging==21.3
    # via
    #   docker
    #   marshmallow
    #   matplotlib
    #   statsmodels
palmerpenguins==0.1.4
    # via -r requirements.in
pandas==1.5.1
    # via
    #   -r requirements.in
    #   flytekit
    #   palmerpenguins
    #   pandas-profiling
    #   phik
    #   seaborn
    #   statsmodels
    #   visions
pandas-profiling==3.4.0
    # via flytekitplugins-deck-standard
patsy==0.5.3
    # via statsmodels
phik==0.12.2
    # via pandas-profiling
pillow==9.2.0
    # via
    #   imagehash
    #   matplotlib
    #   visions
platformdirs==2.5.2
    # via virtualenv
plotly==5.10.0
    # via flytekitplugins-deck-standard
prometheus-client==0.13.1
    # via ray
protobuf==3.20.3
    # via
    #   flyteidl
    #   flytekit
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   protoc-gen-swagger
    #   ray
protoc-gen-swagger==0.1.0
    # via flyteidl
psutil==5.9.3
    # via gpustat
py==1.11.0
    # via retry
py-spy==0.3.14
    # via ray
py4j==0.10.9.5
    # via pyspark
pyarrow==6.0.1
    # via flytekit
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pydantic==1.10.2
    # via pandas-profiling
pyopenssl==22.1.0
    # via flytekit
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
pyspark==3.3.0
    # via
    #   -r requirements.in
    #   flytekitplugins-spark
python-dateutil==2.8.2
    # via
    #   arrow
    #   croniter
    #   flytekit
    #   matplotlib
    #   pandas
python-json-logger==2.0.4
    # via flytekit
python-slugify==6.1.2
    # via cookiecutter
pytimeparse==1.1.8
    # via flytekit
pytz==2022.5
    # via
    #   flytekit
    #   pandas
pywavelets==1.4.1
    # via imagehash
pyyaml==6.0
    # via
    #   cookiecutter
    #   flytekit
    #   pandas-profiling
    #   ray
ray[default]==1.13.0
    # via
    #   -r requirements.in
    #   flytekitplugins-ray
regex==2022.9.13
    # via docker-image-py
requests==2.28.1
    # via
    #   cookiecutter
    #   docker
    #   flytekit
    #   google-api-core
    #   pandas-profiling
    #   ray
    #   responses
responses==0.22.0
    # via flytekit
retry==0.9.2
    # via flytekit
rsa==4.9
    # via google-auth
scikit-learn==1.1.2
    # via -r requirements.in
scipy==1.9.3
    # via
    #   imagehash
    #   missingno
    #   pandas-profiling
    #   phik
    #   scikit-learn
    #   statsmodels
seaborn==0.12.1
    # via
    #   missingno
    #   pandas-profiling
six==1.16.0
    # via
    #   blessed
    #   google-auth
    #   gpustat
    #   grpcio
    #   patsy
    #   python-dateutil
smart-open==6.2.0
    # via ray
sortedcontainers==2.4.0
    # via flytekit
statsd==3.3.0
    # via flytekit
statsmodels==0.13.2
    # via pandas-profiling
tangled-up-in-unicode==0.2.0
    # via visions
tenacity==8.1.0
    # via plotly
text-unidecode==1.3
    # via python-slugify
threadpoolctl==3.1.0
    # via scikit-learn
toml==0.10.2
    # via responses
tqdm==4.64.1
    # via pandas-profiling
types-toml==0.10.8
    # via responses
typing-extensions==4.4.0
    # via
    #   flytekit
    #   pydantic
    #   typing-inspect
typing-inspect==0.8.0
    # via dataclasses-json
urllib3==1.26.12
    # via
    #   docker
    #   flytekit
    #   requests
    #   responses
virtualenv==20.16.5
    # via ray
visions[type_image_path]==0.7.5
    # via pandas-profiling
wcwidth==0.2.5
    # via blessed
websocket-client==1.4.1
    # via docker
wheel==0.37.1
    # via flytekit
wrapt==1.14.1
    # via
    #   deprecated
    #   flytekit
yarl==1.8.1
    # via aiohttp
zipp==3.9.0
    # via importlib-metadata
